+++ src/commander-options.ts
@@ -1,0 +1,29 @@
+import puppeteer = require('puppeteer');
+
+export function commaSeparatedList(
+  value: string,
+  dummyPrevious: any,
+): Array<string> {
+  return value.split(',');
+}
+
+export function generatePuppeteerPDFMargin(
+  value: string,
+  dummyPrevious: any,
+): puppeteer.PDFOptions['margin'] {
+  const marginStrings = value.split(',');
+
+  const marginTop = marginStrings[0];
+  const marginRight = marginStrings[1];
+  const marginBottom = marginStrings[2];
+  const marginLeft = marginStrings[3];
+
+  const generatedMargins = {
+    top: marginTop,
+    right: marginRight,
+    bottom: marginBottom,
+    left: marginLeft,
+  };
+
+  return generatedMargins;
+}
