+++ tsconfig.json
@@ -1,8 +1,10 @@
 {
   "extends": "@tsconfig/node14/tsconfig.json",
   "compilerOptions": {
+    "lib": ["dom"],
     "rootDir": "src",
     "outDir": "lib",
+    "resolveJsonModule": true
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "lib"]
